+++ NEWS.md
@@ -5,11 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
+## Changed
+
+- `delineate()` returns now an S3 object of class `delineation`.
+
 ## Added
 
 - Package-level documentation has been added.
 - Assertions using the `checkmate` package were added to input parameters throughout the package.
 - Examples were expanded to demonstrates the use of all parameters.
+- A `plot()` method was created for objects of class `delineation`.
 
 ## Fixed